Move `DomInteractor` to `interactor` package
