+++ core/mastodon/src/main/java/br/com/orcinus/orca/core/mastodon/instance/registration/webview/dom/interactor/DomInteractor.extensions.kt
@@ -13,7 +13,7 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.core.mastodon.instance.registration.webview.dom
+package br.com.orcinus.orca.core.mastodon.instance.registration.webview.dom.interactor
 
 /**
  * Creates a script from an interaction with a Document Object Model (DOM), made possible by the
